solved the joystick lost package issues, but the joystick speed is very slow

+++ components/joysticks_console/joysticks_console.h
@@ -1,7 +1,7 @@
 #ifndef JOYSTICKS_CONSOLE_H
 #define JOYSTICKS_CONSOLE_H
 
-extern float joystickData[4];  // array to hold joystick data
+extern int8_t joystickData[4];  // array to hold joystick data
 
 // Initialize joystick filters.
 void joysticks_init();
@@ -10,6 +10,6 @@ void joysticks_init();
 void joysticks_read();
 
 // Function to process joystick data and send it to a queue storage
-void processJoystickData(int x1, int y1, int x2, int y2);
+void processJoystickData(int8_t x1, int8_t y1, int8_t x2, int8_t y2);
 
 #endif // JOYSTICKS_CONSOLE_H
